packages/campaign-schema/pyproject.toml
[project]
name = "campaign-schema"
version = "0.1.0"
description = "Shared Pydantic brief + brand config schema for the Adobe creative pipeline"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.11.0",
    "pyyaml>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["campaign_schema"]
